First test of lighting, REALLY (unusably) slow right now.
A block's heightmap is generated when it is loaded (later only when it is first generated)
at which point a ray gets cast down to determine what block is reached by full skylight.
After that lighting updates are scheduled in LightingEngine that deal with scattering light
and light emitting blocks.
These updates are run each game tick